Fix positioning errors with category ticks
Two problems are fixed in this commit:
1) Rounding errors caused by use of int instead of qreal made
   category axes to be one pixel off 50% of the time.
2) Bar category axes offset was calculated incorrectly, and again
   with rounding errors stemming from use of int.

Also fixed some code style issues in related code.

Task-number: QTRD-1920

+++ src/axis/categoryaxis/chartcategoryaxisx.cpp
@@ -53,10 +53,10 @@ QVector<qreal> ChartCategoryAxisX::calculateLayout() const
         qreal scale = gridRect.width() / range;
         for (int i = 0; i < tickCount; ++i) {
             if (i < tickCount - 1) {
-                int x = (m_axis->startValue(m_axis->categoriesLabels().at(i)) - min()) * scale + gridRect.left();
+                qreal x = (m_axis->startValue(m_axis->categoriesLabels().at(i)) - min()) * scale + gridRect.left();
                 points[i] = x;
             } else {
-                int x = (m_axis->endValue(m_axis->categoriesLabels().at(i - 1)) - min())  * scale + gridRect.left();
+                qreal x = (m_axis->endValue(m_axis->categoriesLabels().at(i - 1)) - min())  * scale + gridRect.left();
                 points[i] = x;
             }
         }
